ENH: Add repeat to plan_stubs, refactor count to use repeat
